requirements.txt
Django>=4.2,<5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
django-apscheduler>=0.6
Pillow>=10.0           # for ImageField (profile pictures)
django-environ>=0.10   # optional: env config
requests>=2.28